+++ src/main/java/io/supertokens/webserver/Webserver.java
@@ -265,7 +265,7 @@ public class Webserver extends ResourceDistributor.SingletonResource {
         addAPI(new ConsumeResetPasswordAPI(main));
 
         addAPI(new RequestStatsAPI(main));
-        addAPI(new GetTenantCoreConfigAPI(main));
+        addAPI(new GetTenantCoreConfigForDashboardAPI(main));
 
         StandardContext context = tomcatReference.getContext();
         Tomcat tomcat = tomcatReference.getTomcat();